+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - Fixed the parsing of CodeChef sample cases for certain problems where the input/output headers are not prefixed with "Example" or "Sample"
+- Fixed the parsing of CSES problems on CSES's www subdomain
 
 ## [2.32.0](https://github.com/jmerle/competitive-companion/releases/tag/2.32.0) (2022-02-17)
 - Added a contest parser for Virtual Judge